addming base_test, and other review items
